Update table rows
The real hop count is now shown on the table, instead
of 0 for each hop. Also a hop is not added to the table
if it's lat and lon are the same as the previously added
hop's. This is done so the map icons will not cram into
each other and duplicate hop's do not appear on the
table.